# docs/swagger.yaml
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: REST Shopping Cart
  version: "1.0"
paths:
  /checkout/{userid}:
    get:
      consumes:
      - application/json
      description: InsertToTransaksi
      parameters:
      - description: User Id
        in: path
        minimum: 1
        name: userid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: InsertToTransaksi example
      tags:
      - rest-shoppingcart
  /history/{userid}:
    get:
      consumes:
      - application/json
      description: GetTransaksi
      parameters:
      - description: User Id
        in: path
        minimum: 1
        name: userid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: GetTransaksi example
      tags:
      - rest-shoppingcart
  /history/detail/{transaksiid}:
    get:
      consumes:
      - application/json
      description: DetailTransaksi
      parameters:
      - description: Transaksi Id
        in: path
        minimum: 1
        name: transaksiid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: DetailTransaksi example
      tags:
      - rest-shoppingcart
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"